Put Writer and related stuff to a submodule
